+++ packages/ui/src/index.ts
@@ -36,3 +36,4 @@ export * from './components/ThemeSurface';
 export * from './components/Table';
 export * from './components/DateRangePicker';
 export * from './components/DatePicker';
+export * from './components/MetadataTable';